[rtl] move Bundles in Lane to Bundles.scala
